requirements.txt
ackermann-msgs==2.0.2
action-msgs==2.0.2
action-tutorials-interfaces==0.33.5
action-tutorials-py==0.33.5
actionlib-msgs==5.3.6
actuator-msgs==0.0.1
ament-cmake-test==2.5.4
ament-copyright==0.17.2
ament-cppcheck==0.17.2
ament-cpplint==0.17.2
ament-flake8==0.17.2
ament-index-python==1.8.1
ament-lint==0.17.2
ament-lint-cmake==0.17.2
ament-package==0.16.4
ament-pep257==0.17.2
ament-uncrustify==0.17.2
ament-xmllint==0.17.2
angles==1.16.0
bond==4.1.2
builtin-interfaces==2.0.2
cartographer-ros-msgs==2.0.9003
composition-interfaces==2.0.2
contourpy==1.3.2
control-msgs==5.4.1
controller-manager==4.32.0
controller-manager-msgs==4.32.0
cv-bridge==4.1.0
cycler==0.12.1
demo-nodes-py==0.33.5
diagnostic-msgs==5.3.6
diagnostic-updater==4.2.6
domain-coordinator==0.12.0
dwb-msgs==1.3.7
dynamixel-sdk==3.8.4
example-interfaces==0.12.0
examples-rclpy-executors==0.19.5
examples-rclpy-minimal-action-client==0.19.5
examples-rclpy-minimal-action-server==0.19.5
examples-rclpy-minimal-client==0.19.5
examples-rclpy-minimal-publisher==0.19.5
examples-rclpy-minimal-service==0.19.5
examples-rclpy-minimal-subscriber==0.19.5
filelock==3.18.0
fonttools==4.59.0
fsspec==2025.5.1
generate-parameter-library-py==0.5.0
geographic-msgs==1.0.6
geometry-msgs==5.3.6
gps-msgs==2.1.0
greenlet==3.2.3
image-geometry==4.1.0
interactive-markers==2.5.4
Jinja2==3.1.6
joblib==1.5.1
joint-state-publisher==2.4.0
joint-state-publisher-gui==2.4.0
kiwisolver==1.4.8
laser-geometry==2.7.1
launch==3.4.5
launch-ros==0.26.8
launch-testing==3.4.5
launch-testing-ros==0.26.8
launch-xml==3.4.5
launch-yaml==3.4.5
lifecycle-msgs==2.0.2
logging-demo==0.33.5
map-msgs==2.4.1
MarkupSafe==2.1.5
matplotlib==3.10.3
message-filters==4.11.6
moveit-msgs==2.6.0
mpmath==1.3.0
mystring==0.2.252
narwhals==1.47.1
nav-2d-msgs==1.3.7
nav-msgs==5.3.6
nav2-common==1.3.7
nav2-msgs==1.3.7
nav2-simple-commander==1.0.0
networkx==3.5
numpy==2.2.6
nvidia-cublas-cu12==12.9.1.4
nvidia-cuda-cupti-cu12==12.9.79
nvidia-cuda-nvrtc-cu12==12.9.86
nvidia-cuda-runtime-cu12==12.9.79
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.4.1.4
nvidia-cufile-cu12==1.14.1.1
nvidia-curand-cu12==10.3.10.19
nvidia-cusolver-cu12==11.7.5.82
nvidia-cusparse-cu12==12.5.10.65
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvshmem-cu12==3.3.9
nvidia-nvtx-cu12==12.9.79
object-recognition-msgs==2.0.0
octomap-msgs==2.0.1
opencv-python==4.12.0.88
osrf-pycommon==2.1.6
packaging==25.0
pal-statistics==2.6.2
pal-statistics-msgs==2.6.2
pandas==2.3.1
pandasql==0.7.3
pcl-msgs==1.0.0
pendulum-msgs==0.33.5
pillow==11.3.0
plotly==6.2.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-qt-binding==2.2.1
pytorch-triton==3.4.0+gitae848267
pytz==2025.2
qt-dotgraph==2.7.5
qt-gui==2.7.5
qt-gui-cpp==2.7.5
qt-gui-py-common==2.7.5
quality-of-service-demo-py==0.33.5
rcl-interfaces==2.0.2
rclpy==7.1.4
rcutils==6.7.2
resource-retriever==3.4.4
rmw-dds-common==3.1.0
robot-localization==3.8.2
ros-gz-bridge==1.0.15
ros-gz-interfaces==1.0.15
ros-gz-sim==1.0.15
ros2action==0.32.4
ros2bag==0.26.7
ros2bag-mcap-cli==0.26.7
ros2bag-sqlite3-cli==0.26.7
ros2cli==0.32.4
ros2component==0.32.4
ros2controlcli==4.32.0
ros2doctor==0.32.4
ros2interface==0.32.4
ros2launch==0.26.8
ros2lifecycle==0.32.4
ros2multicast==0.32.4
ros2node==0.32.4
ros2param==0.32.4
ros2pkg==0.32.4
ros2run==0.32.4
ros2service==0.32.4
ros2topic==0.32.4
rosbag2-interfaces==0.26.7
rosbag2-py==0.26.7
rosgraph-msgs==2.0.2
rosidl-adapter==4.6.5
rosidl-cli==4.6.5
rosidl-cmake==4.6.5
rosidl-generator-c==4.6.5
rosidl-generator-cpp==4.6.5
rosidl-generator-py==0.22.1
rosidl-generator-type-description==4.6.5
rosidl-parser==4.6.5
rosidl-pycommon==4.6.5
rosidl-runtime-py==0.13.1
rosidl-typesupport-c==3.2.2
rosidl-typesupport-cpp==3.2.2
rosidl-typesupport-fastrtps-c==3.6.1
rosidl-typesupport-fastrtps-cpp==3.6.1
rosidl-typesupport-introspection-c==4.6.5
rosidl-typesupport-introspection-cpp==4.6.5
rpyutils==0.4.1
rqt-action==2.2.0
rqt-bag==1.5.4
rqt-bag-plugins==1.5.4
rqt-console==2.2.1
rqt-graph==1.5.4
rqt-gui==1.6.0
rqt-gui-py==1.6.0
rqt-msg==1.5.1
rqt-plot==1.4.3
rqt-publisher==1.7.2
rqt-py-common==1.6.0
rqt-py-console==1.2.2
rqt-reconfigure==1.6.2
rqt-service-caller==1.2.1
rqt-shell==1.2.2
rqt-srv==1.2.2
rqt-topic==1.7.3
scikit-learn==1.7.0
scipy==1.16.0
sensor-msgs==5.3.6
sensor-msgs-py==5.3.6
service-msgs==2.0.2
setuptools==78.1.0
shape-msgs==5.3.6
six==1.17.0
slam-toolbox==2.8.3
smclib==4.1.2
SQLAlchemy==2.0.41
srdfdom==2.0.7
sros2==0.13.3
statistics-msgs==2.0.2
std-msgs==5.3.6
std-srvs==5.3.6
stereo-msgs==5.3.6
sympy==1.14.0
teleop-twist-keyboard==2.4.0
tf-transformations==1.1.0
tf2-geometry-msgs==0.36.12
tf2-kdl==0.36.12
tf2-msgs==0.36.12
tf2-py==0.36.12
tf2-ros-py==0.36.12
tf2-sensor-msgs==0.36.12
tf2-tools==0.36.12
threadpoolctl==3.6.0
topic-monitor==0.33.5
torch==2.9.0.dev20250716+cu129
torchaudio==2.8.0.dev20250717+cu129
torchvision==0.24.0.dev20250717+cu129
tqdm==4.67.1
trajectory-msgs==5.3.6
turtlebot3-applications-msgs==1.0.1
turtlebot3-aruco-tracker==1.3.3
turtlebot3-automatic-parking==1.3.3
turtlebot3-automatic-parking-vision==1.3.3
turtlebot3-autorace-camera==1.2.2
turtlebot3-autorace-detect==1.2.2
turtlebot3-autorace-mission==1.2.2
turtlebot3-example==2.3.1
turtlebot3-msgs==2.4.0
turtlebot3-teleop==2.3.1
turtlebot3-yolo-object-detection==1.3.3
turtlesim==1.8.3
type-description-interfaces==2.0.2
typing_extensions==4.14.0
tzdata==2025.2
unique-identifier-msgs==2.5.0
urdfdom-py==1.2.1
vision-msgs==4.1.1
visualization-msgs==5.3.6
xacro==2.0.13
